fix(hyrpland_socket_handling): statically type path lengths

+++ src/utils/hyprland_socket_handling.c
@@ -41,7 +41,8 @@ char *get_hyprland_socket(Socket socket_type) {
     // Add the rest of the path to the hyprland_instance_signature.
     int his_buffer_size = 0;
     his_buffer_size =
-        strlen(xdg_runtime_dir) + strlen(hyprland_instance_signature) + (unsigned long) HIS_PATH_BUFFER_SIZE;
+        (int) strlen(xdg_runtime_dir) + (int) strlen(hyprland_instance_signature)
+        + HIS_PATH_BUFFER_SIZE;
 
     // Configure the socket path depending on the socket given.
     // SOCKET is for "socket" (handles requests); SOCKET2 is for "socket2", (handles events).
